// src/app/page/master/jenis-paket/jenis-paket.interface.ts
import {EventEmitter} from '@angular/core';

export interface JenisPaket {
	
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
 }